added a pagination helper, so i can reuse for connection history
